// src/api/technologyCollection.ts
type TechnologyType =
  | 'JS'
  | 'Java'
  | 'Data'
  | 'Blockchain'
  | 'Html'
  | 'Mobile'
  | 'PHP'
  | 'Python'
  | 'Testing'
  | '.Net'
  | 'C++';

export const techCollection: TechnologyType[] = [
  'JS',
  'Java',
  'Data',
  'Blockchain',
  'Html',
  'Mobile',
  'PHP',
  'Python',
  'Testing',
  '.Net',
  'C++',
];

export default TechnologyType;